Make `TableConfig` and `FeatureConfig` unhashable. (#150)
`TableConfig` contains a keras `Optimizer` and a keras `Initializer`, which are not hashable and even mutable in the case of the optimizer. The "unsafe hash" was actually unsafe. Instead of using `TableConfig`s as keys directly, we use the id of the `TableConfig`, which is correct because we are detecting reused instances.

Also:
- renamed TensorFlow config conversion functions to be more consistent with the JAX ones
- added unit tests for the TensorFlow config conversion functions
- fix optimizer conversion with the Torch backend