Use the JTE API to set stacked table stats to the maximum of the input table specs.
This allows setting parameters like `max_ids_per_partition` and `max_unique_ids_per_partition`,
`suggested_coo_buffer_size` for stacked tables with auto-stacking.

Although the heuristic may not be optimal, this at least provides a method for directly
setting the values in the stacked tables, and is consistent with the default values
if nothing is set.

Uses the `jax_tpu_embedding` API for future-proofing.